Reorganize e2e test site into a unified Astro site with multiple configs
