+++ .gitignore
@@ -4,6 +4,7 @@ node_modules/
 # Build outputs
 dist/
 .astro/
+.cache/
 .wrangler/
 
 # Environment files
