+++ docs/package.json
@@ -18,7 +18,7 @@
 	"dependencies": {
 		"@astrojs/cloudflare": "^14.1.3",
 		"@astrojs/starlight": "^0.41.3",
-		"astro": "^7.0.8",
+		"astro": "^7.1.6",
 		"astro-lilypond": "workspace:*",
 		"starlight-theme-flexoki": "^0.2.3",
 		"wrangler": "^4.112.0"
