+++ e2e/astro.config.satteri.mjs
@@ -1,16 +1,17 @@
-import { unified } from "@astrojs/markdown-remark";
+import mdx from "@astrojs/mdx";
 import { defineConfig } from "astro/config";
 import lilypond from "astro-lilypond";
 
 export default defineConfig({
-	markdown: {
-		processor: unified(),
-	},
+	outDir: "./dist/satteri",
+	cacheDir: "./.cache/satteri",
 	integrations: [
 		lilypond({
 			defaults: {
 				version: "2.26.0",
+				cropScale: 2,
 			},
 		}),
+		mdx(),
 	],
 });
