+++ e2e/astro.config.unified.mjs
@@ -1,7 +1,14 @@
+import { unified } from "@astrojs/markdown-remark";
+import mdx from "@astrojs/mdx";
 import { defineConfig } from "astro/config";
 import lilypond from "astro-lilypond";
 
 export default defineConfig({
+	outDir: "./dist/unified",
+	cacheDir: "./.cache/unified",
+	markdown: {
+		processor: unified(),
+	},
 	integrations: [
 		lilypond({
 			defaults: {
@@ -9,5 +16,6 @@ export default defineConfig({
 				cropScale: 2,
 			},
 		}),
+		mdx(),
 	],
 });
