--- e2e/general/src/pages/index.astro
@@ -1,7 +1,0 @@
----
-import BaseLayout from "../../../shared/BaseLayout.astro";
----
-
-<BaseLayout title="astro-lilypond e2e fixtures (general)">
-	<h1>astro-lilypond e2e fixtures (general)</h1>
-</BaseLayout>
