+++ e2e/package.json
@@ -4,12 +4,11 @@
 	"private": true,
 	"type": "module",
 	"scripts": {
-		"dev:general": "pnpm --filter astro-lilypond build && astro dev --root general",
-		"dev:satteri": "pnpm --filter astro-lilypond build && astro dev --root satteri",
-		"dev:unified": "pnpm --filter astro-lilypond build && astro dev --root unified",
-		"test": "playwright test",
-		"test:ui": "playwright test --ui",
-		"test:headed": "playwright test --headed"
+		"build:lib": "pnpm --filter astro-lilypond build",
+		"dev:satteri": "pnpm run build:lib && astro dev --config astro.config.satteri.mjs",
+		"dev:unified": "pnpm run build:lib && astro dev --config astro.config.unified.mjs",
+		"pretest": "pnpm run build:lib && astro build --config astro.config.satteri.mjs && astro build --config astro.config.unified.mjs",
+		"test": "playwright test"
 	},
 	"dependencies": {
 		"@astrojs/markdown-remark": "^7.2.1",
