--- e2e/satteri/astro.config.mjs
@@ -1,14 +1,0 @@
-import mdx from "@astrojs/mdx";
-import { defineConfig } from "astro/config";
-import lilypond from "astro-lilypond";
-
-export default defineConfig({
-	integrations: [
-		lilypond({
-			defaults: {
-				version: "2.26.0",
-			},
-		}),
-		mdx(),
-	],
-});
