--- e2e/satteri/tsconfig.json
@@ -1,5 +1,0 @@
-{
-	"extends": "astro/tsconfigs/strictest",
-	"include": [".astro/types.d.ts", "**/*"],
-	"exclude": ["dist"]
-}
