+++ e2e/src/pages/collection.astro
@@ -1,7 +1,7 @@
 ---
 import { getEntry } from "astro:content";
 import { Score } from "astro-lilypond";
-import BaseLayout from "../../../shared/BaseLayout.astro";
+import BaseLayout from "../layouts/BaseLayout.astro";
 
 const entry = await getEntry("scores", "example");
 ---
