+++ e2e/src/pages/index.astro
@@ -1,5 +1,5 @@
 ---
-import BaseLayout from "../../../shared/BaseLayout.astro";
+import BaseLayout from "../layouts/BaseLayout.astro";
 ---
 
 <BaseLayout title="astro-lilypond e2e fixtures">
