+++ e2e/src/pages/mdx-fence-and-component.mdx
@@ -1,6 +1,6 @@
 import { Score } from "astro-lilypond";
-import BaseLayout from "../../../shared/BaseLayout.astro";
-import simple from "../../../general/src/scores/simple.ly";
+import BaseLayout from "../layouts/BaseLayout.astro";
+import simple from "../scores/simple.ly";
 
 <BaseLayout title="MDX fence and Score component">
