+++ e2e/src/pages/score-cropped.astro
@@ -1,6 +1,6 @@
 ---
 import { render } from "astro-lilypond";
-import BaseLayout from "../../../shared/BaseLayout.astro";
+import BaseLayout from "../layouts/BaseLayout.astro";
 import multiPage from "../scores/multi-page.ly";
 
 const { Score } = await render(multiPage, { crop: true });
