+++ e2e/src/pages/score-metadata.astro
@@ -1,5 +1,5 @@
 ---
-import BaseLayout from "../../../shared/BaseLayout.astro";
+import BaseLayout from "../layouts/BaseLayout.astro";
 import score from "../scores/metadata.ly";
 
 const { meta } = score;
