+++ e2e/src/pages/score-multipage.astro
@@ -1,6 +1,6 @@
 ---
 import { Score } from "astro-lilypond";
-import BaseLayout from "../../../shared/BaseLayout.astro";
+import BaseLayout from "../layouts/BaseLayout.astro";
 import multiPage from "../scores/multi-page.ly";
 ---
