+++ e2e/src/pages/score-styled.astro
@@ -1,6 +1,6 @@
 ---
 import { Score } from "astro-lilypond";
-import BaseLayout from "../../../shared/BaseLayout.astro";
+import BaseLayout from "../layouts/BaseLayout.astro";
 import multiPage from "../scores/multi-page.ly";
 import simple from "../scores/simple.ly";
 ---
