+++ e2e/src/scores/collection/example.ly
