+++ e2e/src/scores/collection/metadata.ly
