+++ e2e/src/scores/metadata.ly
