+++ e2e/src/scores/simple.ly
