+++ e2e/tests/score-component.spec.ts
