--- e2e/unified/src/pages/fence-lang-tags.md
@@ -1,12 +1,0 @@
----
-title: Fence language tags (unified processor)
-layout: ../../../shared/BaseLayout.astro
----
-
-```lilypond
-\header {
-  title = "Unified Fence Score"
-}
-
-{ c'4 d' e' f' }
-```
