--- e2e/unified/src/pages/index.astro
@@ -1,7 +1,0 @@
----
-import BaseLayout from "../../../shared/BaseLayout.astro";
----
-
-<BaseLayout title="astro-lilypond e2e fixtures (unified processor)">
-	<h1>astro-lilypond e2e fixtures (unified processor)</h1>
-</BaseLayout>
