+++ package.json
@@ -10,7 +10,6 @@
 		"test": "pnpm --filter astro-lilypond test",
 		"test:coverage": "pnpm --filter astro-lilypond test:coverage",
 		"test:e2e": "pnpm --filter e2e test",
-		"test:e2e:ui": "pnpm --filter e2e test:ui",
 		"release": "changeset publish",
 		"check": "biome check",
 		"check:fix": "biome check --write"
