+++ package/package.json
@@ -56,12 +56,12 @@
 		"unist-util-visit": "^5.1.0"
 	},
 	"devDependencies": {
-		"@astrojs/markdown-remark": "^7.2.1",
-		"@astrojs/markdown-satteri": "^0.3.4",
+		"@astrojs/markdown-remark": "^7.2.2",
+		"@astrojs/markdown-satteri": "^0.3.5",
 		"@types/mdast": "^4.0.4",
 		"@types/node": "^26.1.1",
 		"@vitest/coverage-v8": "^4.1.10",
-		"astro": "^7.0.8",
+		"astro": "^7.1.6",
 		"satteri": "^0.9.5",
 		"typescript": "^7.0.2",
 		"unified": "^11.0.5",
